wake_gestures: accept 2 as a valid input for dt2w
I am so sick of users complaining that KA doesn't work with fullscreen, which is the intuitive option for our device as we only have two states (on or off), as opposed to some devices that have three states (off, on for half of the screen, or on for the entire screen). Make it so that if the user selects
fullscreen (echoing 2 to the path) that it still works.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>